Shit should work now.
Play with it and let me know if you find any bugs.  If you could do a
few QoL edits, that would be great.  Removing commented/uncommented
logs, changing the 1s and 5s on the board into X's and O's, leaving the
blank spots blank, changing the user input to correspond to the visual
row, not the array index, stuff like that.